backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1

# AI Model / NLP
torch==2.1.2
transformers==4.45.1
sentence-transformers==3.0.1

# Stable Diffusion (Diffusers)
diffusers==0.30.2
safetensors==0.4.2

# Retrieval (RAG System)
faiss-cpu==1.8.0
rank-bm25==0.2.2

# GPT OpenAI API
openai==1.52.0

# Image handling
Pillow==10.3.0

# Utilities
numpy==1.26.4
pydantic==2.7.4

# File uploads / CORS
python-multipart==0.0.9
starlette==0.41.2
